% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_list_to_plates}
\alias{add_list_to_plates}
\title{Add a property list to a list of plates}
\usage{
add_list_to_plates(plates, id, list)
}
\arguments{
\item{plates}{A list of plates}

\item{id}{The id of the property of the list you are adding}

\item{list}{The list of values of specific property of well}
}
\value{
A plate list with added property for each well
}
\description{
\code{create_plates} creates a new plate list using specific property as guideline.
}
