% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_lists_to_plates}
\alias{add_lists_to_plates}
\title{Add a list of property vectors to a list of plates}
\usage{
add_lists_to_plates(plates, id, lists)
}
\arguments{
\item{plates}{A list of plates}

\item{id}{The id of the property of the list you are adding}

\item{lists}{The list of vectors of specific property of well}
}
\value{
A plate list with added property for each well
}
\description{
\code{add_list_to_plates} adds list of property vectors to respective plates in the
given list of plates.
}
