% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_constructs}
\alias{create_constructs}
\title{Create construct ids}
\usage{
create_constructs(tgene_f, tgene_fo, WRITE_CONSTRUCT_FILE)
}
\arguments{
\item{tgene_f}{A construct file path}

\item{tgene_fo}{A file path to write modified construct df to}
}
\value{
The data frame read from \code{tgene_f} with added `construct` column
}
\description{
\code{create_constructs} reads construct file and creates unique construct id column.
}
