% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_new_plate_from_list}
\alias{create_new_plate_from_list}
\title{Create new plate from a list}
\usage{
create_new_plate_from_list(well_alpha, well_numer, id, list)
}
\arguments{
\item{well_alpha}{A vector of letters from the alphabet corresponding to
barcode plate rows}

\item{well_numer}{A vector of numbers corresponding to barcode plate columns}

\item{id}{The id of the property of the list you are creating plate from}

\item{list}{The list of values of specific property of well}
}
\value{
A plate list
}
\description{
\code{create_new_plate_from_list} creates a new plate list using specific property as
guideline.
}
