% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_plates}
\alias{create_plates}
\title{Create plates}
\usage{
create_plates(plate_ids, barcode_maps)
}
\arguments{
\item{plate_ids}{A vector of plate id's}

\item{barcode_maps}{A vector of barcode id's that has the same length as \code{plate_ids}.}
}
\value{
A list of the plate lists that were created.
}
\description{
\code{create_plates} creates a plate list for each \code{plate_id} using appropriate barcode maps.
}
