% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cell_qualities}
\alias{get_cell_qualities}
\title{Get cell qualities}
\usage{
get_cell_qualities(plate_ids, cell_quals_f)
}
\arguments{
\item{plate_ids}{A vector of plate id's}

\item{cell_quals_f}{A vector of file paths to cell quality .xlsx files}
}
\value{
A list of vectors of cell qualities
}
\description{
\code{get_cell_qualities} reads multiple cell quality files and returns list of vectors
}
