% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_constructs}
\alias{get_constructs}
\title{Get construct ids}
\usage{
get_constructs(plate_ids, constructs_f, constructs_fo, WRITE_CONSTRUCT_FILE)
}
\arguments{
\item{plate_ids}{A vector of plate id's}

\item{constructs_f}{A vector of file paths to constructs .xlsx files}

\item{constructs_fo}{A vector of file paths to write modified constructs .xlsx files}
}
\value{
A list of vectors of constructs
}
\description{
\code{get_constructs} reads construct file and creates unique construct id
column.
}
