% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_info_from_platename}
\alias{get_info_from_platename}
\title{Get information from plate name}
\usage{
get_info_from_platename(concat_str, id = c("construct_map", "day",
  "replicate"))
}
\arguments{
\item{concat_str}{A concatenated plate id (e.g., "DBI31_D12_R3")}

\item{id}{The property to get from \code{concat_str}. In the format of
{construct_map}_{day}_{replicate}.}
}
\value{
A string of the property value
}
\description{
\code{get_info_from_platename} gets specified information from the
concatenated id of a plate (e.g., "DBI31_D12_R3")
}
