% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make project directory structure}
\usage{
make_project(project_dir, pools, overwrite = TRUE)
}
\arguments{
\item{project_dir}{A file path}

\item{pools}{A vector of pool ID's}

\item{overwrite}{A boolean that if true will overwrite any existing project 
directory specified at \code{project_dir}}
}
\value{
NA
}
\description{
\code{traverse_into_plate_wells} traverses through a plate list and sets
\code{input_data} to property \code{id} for each well.
}
\details{
Running this helper function will generate a project directory
structure given the proper parameters. The folder hierarchy that 
is generated will look like this:

\itemize{
  \item \{project_run_name\}/
  \itemize{
    \item barcodes/
    \item cell_viabilities/
    \itemize{
      \item \{project_run_name\}_P\{\code{pools[1]}\}/
      \item \{project_run_name\}_P\{\code{pools[2]}\}/
      \item ...
    }
    \item config/
    \item constructs/
    \item fastq/
    \itemize{
      \item \{project_run_name\}_P\{\code{pools[1]}\}/
        \itemize{
          \item \{project_run_name\}_P\{\code{pools[1]}\}_R1.fastq
          \item \{project_run_name\}_P\{\code{pools[1]}\}_R2.fastq
        }
      \item \{project_run_name\}_P\{\code{pools[2]}\}/
        \itemize{
          \item ...
        }
      \item ...
    }
    \item logs/
    \item output/
    \item sam/
    \itemize{
      \item \{project_run_name\}_P\{\code{pools[1]}\}/
      \item \{project_run_name\}_P\{\code{pools[2]}\}/
      \item ...
    }
    \item src/
  }
}
}
