% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_s1}
\alias{run_s1}
\title{Run step 1 - Alignment}
\usage{
run_s1(yaml_fo, doS1 = TRUE)
}
\arguments{
\item{yaml_fo}{A file path to the configuration yaml}

\item{doS1}{A boolean that decides whether alignment should be done or not,
default is TRUE}
}
\value{
NA
}
\description{
\code{runs_s1} runs s1.sh and passes in the configuration yaml file
}
