% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_s2}
\alias{run_s2}
\title{Run step 2 - Alignment}
\usage{
run_s2(yaml_fo, doS2 = TRUE)
}
\arguments{
\item{yaml_fo}{A file path to the configuration yaml}

\item{doS2}{A boolean that decides whether alignment should be done or not,
default is TRUE}
}
\value{
NA
}
\description{
\code{runs_s2} runs s2.sh and passes in the configuration yaml file
}
