% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{traverse_into_plate_wells}
\alias{traverse_into_plate_wells}
\title{Traverse into plate wells}
\usage{
traverse_into_plate_wells(plate, id, input_data)
}
\arguments{
\item{plate}{A plate list}

\item{id}{A property of a well}

\item{input_data}{A list of values to be set}
}
\value{
The modified plate list
}
\description{
\code{traverse_into_plate_wells} traverses through a plate list and sets \code{input_data} to property \code{id} for each well.
}
