% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{traverse_outof_plate_wells}
\alias{traverse_outof_plate_wells}
\title{Traverse out of plate wells}
\usage{
traverse_outof_plate_wells(plate, id)
}
\arguments{
\item{plate}{A plate list}

\item{id}{A property of a well}
}
\value{
A vector of values of property \code{id} across all wells
}
\description{
\code{traverse_outof_plate_wells} traverses through a plate list and gets
value at property \code{id} for each well.
}
