% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_wells_info}
\alias{write_wells_info}
\title{Write list of plates to yaml file}
\usage{
write_wells_info(output_dir, plates, WRITE_WELLS_FILE, WRITE_PRINTSHEET_HELPER)
}
\arguments{
\item{output_dir}{A path to directory to output files to}

\item{plates}{A list of plate lists that carries all the well information}

\item{WRITE_WELLS_FILE}{If \code{TRUE}, write .yaml file, otherwise don't}

\item{WRITE_PRINTSHEET_HELPER}{If \code{TRUE}, create pipetting helper file, otherwise don't}
}
\value{
NA
}
\description{
\code{create_plates} creates a new plate list using specific property as guideline.
}
