% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make project directory structure}
\usage{
make_project(output_dir, plate_name, overwrite = TRUE)
}
\arguments{
\item{output_dir}{A file path of where the plate folder will be created}

\item{plate_name}{A string of the plate name}

\item{overwrite}{A boolean that if true will overwrite any existing project 
directory specified at \code{project_dir}}
}
\value{
The project_run_dir path
}
\description{
\code{traverse_into_plate_wells} traverses through a plate list and sets
\code{input_data} to property \code{id} for each well.
}
\details{
Running this helper function will generate a project directory
structure given the proper parameters. The function will also copy the 
necessary scripts into the src folder. The entire folder hierarchy that 
is generated will look like this:

\itemize{
  \item \{plate_name\}/
  \itemize{
    \item cell_viabilities/
    \item config/
    \item logs/
    \item outputs/
    \itemize{
      \item sam/
    }
    \item src/
    \itemize{
      \item perl/
      \item sh/
    }
  }
}
}
